simple_backend/pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "simple-backend"
version = "0.1.0"
description = "A simple task tracker backend with FastAPI and AI integration"
authors = ["Igor Gorelov <ney24fiz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = "^0.111.0"
requests = "^2.31.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.29.0"}
